+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## 0.0.2
 
 - Fix memory metrics in on linux (#6)
+- max_workers variable is now max_worker_limit